update binding to newest ESH/OH version, fix linkRegistry bug

+++ addons/binding/org.openhab.binding.ebus/src/main/java/org/openhab/binding/ebus/internal/EBusLibClient.java
@@ -137,7 +137,7 @@ public class EBusLibClient {
 
         if (command instanceof State) {
             State state = (State) command;
-            DecimalType decimalValue = (DecimalType) state.as(DecimalType.class);
+            DecimalType decimalValue = state.as(DecimalType.class);
 
             if (decimalValue == null) {
                 decimalValue = new DecimalType(state.toString());
